requirements.txt
asyncio
aiohttp
asyncpg
asyncpg_stubs
cachetools
discord.py
jishaku @ git+https://github.com/Gorialis/jishaku
python-dotenv
typing_extensions
gitpython
psutil
humanfriendly